# api/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CurrencyBalanceAfterWithdraw:
    properties:
      EUR:
        description: |-
          Balance in EUR
          default: 50.0
        type: number
      RUB:
        description: |-
          Balance in RUB
          default: 5000.0
        type: number
      USD:
        description: |-
          Balance in USD
          default: 100.0
        type: number
    type: object
...
  handlers.ExchangeRates:
    properties:
      EUR:
        description: |-
          EUR exchange rate
          default: 0.85
        type: number
      RUB:
        description: |-
          RUB exchange rate
          default: 90.0
        type: number
      USD:
        description: |-
          USD exchange rate
          default: 1.0
        type: number
    type: object
  handlers.ExchangeRatesErrorResponse:
    properties:
      error:
        description: |-
          Error message
          default: Failed to retrieve exchange rates
        type: string
    type: object
  handlers.ExchangeRatesResponse:
    properties:
      rates:
        allOf:
        - $ref: '#/definitions/handlers.ExchangeRates'
        description: Exchange rates
    type: object
...
  handlers.WithdrawErrorResponse:
    properties:
      error:
        description: |-
          Error message
          default: Insufficient funds or invalid amount
        type: string
    type: object
  handlers.WithdrawRequest:
    properties:
      amount:
        description: |-
          Amount to withdraw
          required: true
          default: 50.0
        type: number
      currency:
        description: |-
          Currency
          required: true
          default: USD
        type: string
    type: object
  handlers.WithdrawResponse:
    properties:
      message:
        description: |-
          Success message
          default: Withdrawal successful
        type: string
      new_balance:
        allOf:
        - $ref: '#/definitions/handlers.CurrencyBalanceAfterWithdraw'
        description: New balance of the user
    type: object
host: localhost:8080
info:
  contact: {}
  description: Microservice for managing user wallets and currency exchange
  title: gw-currency-wallet API
  version: 1.0.0
paths:
...
  /exchange/rates:
    get:
      description: Fetches current exchange rates for all supported currencies
      produces:
      - application/json
      responses:
        "200":
          description: Exchange rates
          schema:
            $ref: '#/definitions/handlers.ExchangeRatesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ExchangeRatesErrorResponse'
        "500":
          description: Failed to retrieve exchange rates
          schema:
            $ref: '#/definitions/handlers.ExchangeRatesErrorResponse'
      security:
      - BearerAuth: []
      summary: Get exchange rates
      tags:
      - exchange
...
  /wallet/withdraw:
    post:
      consumes:
      - application/json
      description: Withdraw funds from user wallet. Validates amount and currency.
        Checks for sufficient funds.
      parameters:
      - description: Withdraw Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.WithdrawRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Withdrawal successful
          schema:
            $ref: '#/definitions/handlers.WithdrawResponse'
        "400":
          description: Insufficient funds or invalid amount
          schema:
            $ref: '#/definitions/handlers.WithdrawErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.WithdrawErrorResponse'
      security:
      - BearerAuth: []
      summary: Withdraw funds
      tags:
      - wallet
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"